Rename the CamelCase modules and packages to lower case
PEP 8 reserves CamelCase for classes and asks for lower_case module and
package names. RESKit used CamelCase for both, so a name such as NCSource
meant either the module or the class depending on which one the reader
happened to be looking at, and reaching the class through the package
namespace was ambiguous enough to break the whole import graph once the
imports in reskit/weather/__init__.py were reordered.

Renamed, with every call site updated:

  reskit/util/create_LRA.py            -> long_run_average.py
  reskit/weather/NCSource.py           -> nc_source.py
  reskit/weather/SarahSource.py        -> sarah_source.py
  reskit/weather/CordexSource.py       -> cordex_source.py
  reskit/weather/CosmoSource/          -> cosmo_source/
  reskit/weather/Era5Source/           -> era5_source/
      Era5Source.py                    -> era5_source.py
      Era5ZarrSource.py                -> era5_zarr_source.py
      Era5Prepare.py                   -> era5_prepare.py
  reskit/weather/MerraSource/          -> merra_source/
      MerraSource.py                   -> merra_source.py
  reskit/weather/IconlamSource/        -> iconlam_source/
      IconlamSource.py                 -> iconlam_source.py
  reskit/weather/GSA_mean/             -> gsa_mean/
      GSAmeanSource.py                 -> gsa_mean_source.py
  reskit/weather/GWA_mean/             -> gwa_mean/
      GWAmeanSource.py                 -> gwa_mean_source.py

The classes keep their CamelCase names, so the documented API is unchanged:
rk.weather.Era5Source is still the class, rk.weather.Era5Source.LONG_RUN_AVERAGE_GHI_2020_03
still reads a class attribute, and the notebooks need no edit.

create_LRA becomes long_run_average rather than create_lra so that the module
is not named after one of its functions, and so that the rename is not a
case-only one on a case-insensitive filesystem.

The old import paths stop working in this commit. The next commit restores
them as deprecation aliases.
