+++ reskit/__init__.py
@@ -8,5 +8,5 @@ from reskit.workflow_manager import (
     execute_workflow_iteratively,
     load_workflow_result,
 )
-from reskit.weather.Era5Source.Era5Prepare import prepare_era5
+from reskit.weather.era5_source.era5_prepare import prepare_era5
 from reskit.util.input_preparation import download_and_process
