+++ reskit/util/__init__.py
@@ -10,4 +10,4 @@ from reskit.util.topography import visibility_from_topography
 from reskit.util.weather_tile import get_dataframe_with_weather_tilepaths
 
 # ERA5 long-run averages (LRA)
-from reskit.util.create_LRA import create_LRA, create_DNI_LRA
+from reskit.util.long_run_average import create_LRA, create_DNI_LRA
