--- reskit/weather/CosmoSource/__init__.py
@@ -1,1 +1,0 @@
-from .CosmoSource import CosmoSource
