--- reskit/weather/Era5Source/__init__.py
@@ -1,2 +1,0 @@
-from .Era5Source import Era5Source
-from .Era5ZarrSource import Era5ZarrSource
