--- reskit/weather/GWA_mean/__init__.py
@@ -1,1 +1,0 @@
-from .GWAmeanSource import GWAmeanSource
