--- reskit/weather/MerraSource/__init__.py
@@ -1,1 +1,0 @@
-from .MerraSource import MerraSource
