+++ reskit/weather/__init__.py
@@ -1,13 +1,13 @@
 from . import util
 
-from .NCSource import NCSource
+from .nc_source import NCSource
 
-from .MerraSource import MerraSource
-from .Era5Source import Era5Source, Era5ZarrSource
-from .SarahSource import SarahSource
-from .GSA_mean import GSAmeanSource
-from .GWA_mean import GWAmeanSource
-from .IconlamSource import IconlamSource
+from .merra_source import MerraSource
+from .era5_source import Era5Source, Era5ZarrSource
+from .sarah_source import SarahSource
+from .gsa_mean import GSAmeanSource
+from .gwa_mean import GWAmeanSource
+from .iconlam_source import IconlamSource
 
-# from .CosmoSource import CosmoSource  # TODO: NEEDS UPDATING!
+# from .cosmo_source import CosmoSource  # TODO: NEEDS UPDATING!
 # from .CordexSource import CordexSourc  # TODO: NEEDS UPDATING!
