+++ reskit/weather/cosmo_source/__init__.py
@@ -1,0 +1,1 @@
+from .cosmo_source import CosmoSource
