+++ reskit/weather/cosmo_source/cosmo_source.py
@@ -2,7 +2,7 @@
 
 import pytz
 
-from ..NCSource import *
+from ..nc_source import *
 
 # Define constants
