+++ reskit/weather/era5_source/data/ERA5_100m_wind_direction.processed_2008_2017_mean_global_2026_07.tiff
