+++ reskit/weather/era5_source/data/ERA5_DNI_1994_2018_mean_global_2026_07.tiff
