+++ reskit/weather/era5_source/data/ERA5_DNI_mean.tif
