+++ reskit/weather/era5_source/data/ERA5_surface_solar_radiation_downwards.processed.t_adjusted_1994_2018_mean_global_2026_07.tiff
