+++ reskit/weather/era5_source/data/ERA5_total_sky_direct_solar_radiation_at_surface.processed.t_adjusted_1994_2018_mean_global_2026_07.tiff
