+++ reskit/weather/era5_source/data/ERA5_wind_speed_100m_mean.tiff
