+++ reskit/weather/gsa_mean/GSA_DNI_mean_ERA5_pixel.tif
