+++ reskit/weather/gsa_mean/GSA_GHI_mean_ERA5_pixel.tif
