+++ reskit/weather/gsa_mean/gsa_mean_source.py
@@ -1,6 +1,6 @@
 from os.path import dirname, join
 
-from ..NCSource import NCSource
+from ..nc_source import NCSource
 
 
 class GSAmeanSource:
