+++ reskit/weather/gwa_mean/GWA_mean_ERA5_pixel.tif
