+++ reskit/weather/gwa_mean/__init__.py
@@ -1,0 +1,1 @@
+from .gwa_mean_source import GWAmeanSource
