+++ reskit/weather/merra_source/data/gwa50_mean_over_merra.tif
