+++ reskit/weather/merra_source/data/merra_average_SWGDN_1994-2015_globe.tif
