+++ reskit/weather/merra_source/merra_source.py
@@ -3,7 +3,7 @@ from os.path import dirname, join
 import geokit as gk
 import numpy as np
 
-from ..NCSource import NCSource
+from ..nc_source import NCSource
 
 
 class MerraSource(NCSource):
